+++ modules/HyperIoTHProject/HyperIoTHProject-model/src/main/java/it/acsoftware/hyperiot/hpacket/model/HPacketTrafficPlan.java
@@ -19,16 +19,21 @@ package it.acsoftware.hyperiot.hpacket.model;
 
 public enum HPacketTrafficPlan {
 
-    LOW("low"), MEDIUM("medium"), HIGH("high"), INTENSIVE("intensive");
+    LOW("low", 1), MEDIUM("medium", 2), HIGH("high", 3), INTENSIVE("intensive", 4);
 
     private String name;
+    private int order;
 
-    HPacketTrafficPlan(String name) {
+    HPacketTrafficPlan(String name, int order) {
         this.name = name;
+        this.order = order;
     }
 
     public String getName() {
         return name;
     }
 
+    public int getOrder() {
+        return order;
+    }
 }