+++ ReleaseNotes.md
@@ -8,3 +8,4 @@
 - Bug: Added png format to area image
 - Update: Added default area view type - IMAGE
 - Update: Added /hyperiot/services/version end point which return current services version
+- Update: Adding new column family HBase in order to support file streaming and store them inside HBase